requirements.txt
alembic==1.13.3
appnope==0.1.4
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boto3==1.35.5
botocore==1.35.5
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cssselect2==0.8.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
dominate==2.9.1
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.1
environs==11.0.0
Events==0.5
executing==2.1.0
fastjsonschema==2.20.0
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Bootstrap4==4.0.2
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.54.1
google-generativeai==0.3.2
greenlet==3.0.3
gunicorn==23.0.0
html5lib==1.1
idna==3.8
ipython==8.12.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.3.0
Mako==1.3.5
MarkupSafe==3.0.1
marshmallow==3.22.0
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
opensearch-dsl==2.1.0
opensearch-py==2.7.1
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pgsync==3.2.1
pickleshare==0.7.5
pillow==10.4.0
pip-check-reqs==2.5.3
pipreqs==0.5.0
platformdirs==4.3.6
postgres==4.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydyf==0.11.0
Pygments==2.18.0
pyotp==2.9.0
pypandoc==1.14
pyphen==0.16.0
openai==1.40.2
httpx==0.27.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.2
pyzmq==26.2.0
qrcode==8.0
redis==5.0.8
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
requests-aws4auth==1.3.1
rpds-py==0.20.0
s3transfer==0.10.2
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.32
sqlparse==0.5.1
stack-data==0.6.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==1.26.19
visitor==0.1.3
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.0.4
WTForms==3.1.2
WTForms-SQLAlchemy==0.4.1
yarg==0.1.9
zopfli==0.2.3
